Refactored test cases to use common function
In a lot of test cases we are using the `socket`/`bind`/`listen`
sequence along with port numbers. This was often causing `EADDRINUSE`
error and also, this was a lot of duplication of the code. Now we use a
common function and a global port number during testing.

Also removed a test case from `connectx` on unconnected socket, this is
being handled as part of `accept` test case.