Implementation of `shutdown` and some API cleanup
No longer provide an API called `close` on any of the sockets, instead,
`drop` on `SctpListener` and `SctpConnectedSocket` now calls the `close`
call for the `inner` file descriptor.

We could as well have implemnted `SctpListener` and
`SctpConnectedSocket` using `OwnedFd` as inner and then this is not
explicitly required to be done as `OwnedFd` does call the `close` on the
file descriptor, but that would restrict us to Rust versions 1.63.0 or
higher, so for now we just do it ourselves.