Cleanup of the way internal functions are called.
Originally, our `inner` structure was a `RawFd` and hence all the
internal functions were using this `RawFd` as a parameter. However, once
that is converted to `AsyncFd`, at a number of places we had this ugly
`*self.inner.get_ref()` while making the call. Now consistently pass
`&self.inner` (or `self.inner` where it is to be consumed) to internal
API calls. This makes writing unit tests for `internal` functions is
easier.