Fix: `sctp_send_internal` when `SendInfo` not None
When `SendInfo` is set to `Some` value, we were not copying the contents
of the `SendInfo` to the control message, but the control message was
still non null. This could in worst case result in a `Shutdown` being
sent on the association. Now properly using the value of the `SendInfo`
passed to the function.

Also added test case for it.