Refactored rust actions to custom actions
Refactored rust build and test actions to a custom action, so that both
these actions can be run on different runners using the same code.

Updated the build jobs to run on ubuntu as well as latest apple.

Do not run `cargo test` on MacOs the builds are failing right now.

Fixed warnings and removed stray `eprintln!`